requirements.txt
telethon
discord.py
python-dotenv
qrcode[pil]
aiohttp
watchfiles
ruff
ty
